separating fake communicator from mpifile
