+++ pixell/mpi.py
@@ -2,28 +2,7 @@
 from __future__ import print_function
 import sys, os, traceback, copy, numpy as np
 
-def _unbuf(bufspec):
-    return bufspec[0] if isinstance(bufspec, tuple) else np.asarray(bufspec)
-
-class FakeCommunicator:
-    def __init__(self):
-        self.size = 1
-        self.rank = 0
-    def Allreduce(self, sendbuf, recvbuf, op=lambda a,b:a+b):
-        _unbuf(recvbuf)[()] = _unbuf(sendbuf)
-    def Allgather(self, sendbuf, recvbuf):
-        _unbuf(recvbuf)[0] = _unbuf(sendbuf)
-    def Allgatherv(self, sendbuf, redvbuf):
-        _unbuf(recvbuf)[()] = _unbuf(sendbuf)
-    def Alltoallv(self, sendbuf, recvbuf):
-        _unbuf(recvbuf)[()] = _unbuf(sendbuf)
-    def Barrier(self): pass
-    def allreduce(self, sendobj, op=lambda a,b:a+b):
-        return copy.deepcopy(sendobj)
-    def allgather(self, sendobj, op=lambda a,b:a+b):
-        return [copy.deepcopy(sendobj)]
-    def allgatherv(self, sendobj, op=lambda a,b:a+b):
-        return [copy.deepcopy(sendobj)]
+from .fake_communicator import FakeCommunicator
 
 FAKE_WORLD = FakeCommunicator()
 COMM_WORLD = FAKE_WORLD
